refactor: refactor revolution page into modular components
Split the revolution page into reusable components: RevolutionContent, RevolutionStepCard, ScrollToTopButton, and VideoPlayer. Moved revolutionSteps data to a constants file. Simplified page.tsx by removing inline logic and markup, delegating UI and behavior to new components for better maintainability and readability.
