+++ src/app/cach-mang-giai-phong-dan-toc/constants/revolution-steps.ts
@@ -1,0 +1,77 @@
+export const revolutionSteps = [
+  {
+    id: 1,
+    title: "Con đường cách mạng vô sản",
+    subtitle: "Muốn cứu nước và giải phóng dân tộc",
+    content: [
+      "Từ thực tiễn phong trào yêu nước nổ ra và thất bại",
+      "Khảo cứu cách mạng Pháp, Mỹ, đặc biệt là thắng lợi của Cách mạng Tháng Mười Nga năm 1917",
+      "Năm 1920, khi đọc bản Sơ thảo lần thứ nhất những luận cương về vấn đề dân tộc và vấn đề thuộc địa của Lênin",
+      "Kết luận: Độc lập dân tộc gắn liền với chủ nghĩa xã hội"
+    ],
+    quote: "Muốn cứu nước và giải phóng dân tộc không có con đường nào khác con đường cách mạng vô sản",
+    image: "/images/giai-phong-dan-toc/cach-mang-vo-san.jpg",
+    altText: "Hồ Chí Minh và con đường cách mạng",
+    color: "from-red-800 to-red-600"
+  },
+  {
+    id: 2,
+    title: "Đảng Cộng sản lãnh đạo",
+    subtitle: "Nhân tố chủ quan quyết định",
+    content: [
+      "Đảng Cộng sản là nhân tố chủ quan để giai cấp công nhân hoàn thành sứ mệnh lịch sử",
+      "Giai cấp công nhân phải tổ chức ra chính đảng",
+      "Đảng phải thuyết phục, giác ngộ và tập hợp đồng bào quần chúng",
+      "Đảng là đội tiên phong của giai cấp công nhân và nhân dân lao động"
+    ],
+    quote: "Đảng của cả dân tộc Việt Nam",
+    image: "/images/giai-phong-dan-toc/dang-cong-san-viet-nam-lanh-dao.jpg",
+    altText: "Biểu tượng lãnh đạo của Đảng",
+    color: "from-yellow-600 to-orange-500"
+  },
+  {
+    id: 3,
+    title: "Đại đoàn kết toàn dân tộc",
+    subtitle: "Lấy liên minh công - nông làm nền tảng",
+    content: [
+      "Cách mạng là sự nghiệp của quần chúng nhân dân",
+      "Quần chúng nhân dân là chủ thể sáng tạo ra lịch sử",
+      "Có dân là có tất cả, trên đời này không gì quý bằng dân",
+      "Công nông là 'chủ cách mệnh... là gốc cách mệnh'"
+    ],
+    quote: "Được lòng dân thì được tất cả, mất lòng dân thì mất tất cả",
+    image: "/images/giai-phong-dan-toc/dai-doan-ket-dan-toc.jpg",
+    altText: "Đại đoàn kết dân tộc",
+    color: "from-green-700 to-green-500"
+  },
+  {
+    id: 4,
+    title: "Chủ động, sáng tạo",
+    subtitle: "Giành thắng lợi trước cách mạng vô sản ở chính quốc",
+    content: [
+      "Luận điểm sáng tạo của Hồ Chí Minh, bổ sung vào Chủ nghĩa Mác – Lênin",
+      "Thuộc địa có vị trí, vai trò đặc biệt đối với chủ nghĩa đế quốc",
+      "Tinh thần đấu tranh cách mạng hết sức quyết liệt của các dân tộc thuộc địa",
+      "Hình thành một 'lực lượng khổng lồ' khi được tập hợp và giác ngộ"
+    ],
+    quote: "Sáng tạo có giá trị lý luận và thực tiễn",
+    image: "/images/giai-phong-dan-toc/ly-luan-sang-tao.jpg",
+    altText: "Tư duy sáng tạo và chủ động",
+    color: "from-blue-700 to-blue-500"
+  },
+  {
+    id: 5,
+    title: "Bạo lực cách mạng",
+    subtitle: "Phương pháp đấu tranh cần thiết",
+    content: [
+      "Giải phóng dân tộc là một cuộc cách mạng bạo lực",
+      "Sự thống trị thực dân, phát xít, đế quốc không có cơ sở cho đấu tranh phi bạo lực",
+      "Giai cấp thống trị không bao giờ từ bỏ địa vị thống trị",
+      "Bạo lực của quần chúng với hai lực lượng chính trị và quân sự"
+    ],
+    quote: "Dùng bạo lực cách mạng để chống lại bạo lực phản cách mạng",
+    image: "/images/giai-phong-dan-toc/cach-mang-dau-tranh.jpg",
+    altText: "Tinh thần đấu tranh cách mạng",
+    color: "from-purple-700 to-purple-500"
+  }
+];
