+++ src/components/scroll-to-top-button.tsx
@@ -1,0 +1,39 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { ChevronUp } from "lucide-react";
+
+export default function ScrollToTopButton() {
+  const [showScrollToTop, setShowScrollToTop] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setShowScrollToTop(window.scrollY > 500);
+    };
+
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, []);
+
+  const scrollToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: 'smooth'
+    });
+  };
+
+  if (!showScrollToTop) return null;
+
+  return (
+    <button
+      onClick={scrollToTop}
+      className="fixed right-2 bottom-2 sm:right-6 sm:bottom-6 z-50 
+                 bg-red-600 hover:bg-red-700 text-white p-2 sm:p-3 rounded-full
+                 shadow-lg transition-all duration-300 hover:scale-110
+                 animate-bounce"
+      aria-label="Scroll to top"
+    >
+      <ChevronUp className="w-5 h-5 sm:w-6 sm:h-6" />
+    </button>
+  );
+}
